requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
psycopg2-binary
google-generativeai